requirements.txt
fastapi
uvicorn[standard]
redis
fhir.resources==8.0.0
